split dtos so that there is one per file
