+++ src/domain/agent/credential/dto/credential.dto.delete.ts
