+++ src/domain/agent/credential/dto/credential.dto.metadata.ts
