+++ src/domain/agent/credential/index.ts
@@ -1,6 +1,4 @@
 export * from './credential.entity';
 export * from './credential.interface';
-export * from './credential.dto.create';
-export * from './credential.dto.delete';
-export * from './credentials.dto.search';
-export * from './credentials.dto.search';
+export * from './dto/credential.dto.create';
+export * from './dto/credentials.dto.search';
