+++ src/domain/common/authorization-policy/authorization.policy.service.ts
@@ -15,7 +15,7 @@ import {
 import { AuthorizationPolicy } from '@domain/common/authorization-policy/authorization.policy.entity';
 import { IAuthorizationPolicy } from './authorization.policy.interface';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
-import { CredentialsSearchInput } from '@domain/agent/credential/credentials.dto.search';
+import { CredentialsSearchInput } from '@domain/agent/credential/dto/credentials.dto.search';
 import { IAuthorizationPolicyRuleCredential } from '../../../core/authorization/authorization.policy.rule.credential.interface';
 import { AuthorizationPolicyRuleCredential } from '@core/authorization/authorization.policy.rule.credential';
 import { AuthorizationService } from '@core/authorization/authorization.service';
